fix(config): address codex review — same-model fallback and tier/whitelist warn
1. Allow same-model fallback when parent tool is in the whitelist
   (user explicitly listed it = intent to use it, even for adversarial)
2. Warn when tier and whitelist have no intersection instead of
   silent fallthrough (makes the configuration conflict visible)
3. Extract debate resolve logic to debate_cmd_resolve.rs (monolith limit)
