+++ crates/cli-sub-agent/src/debate_cmd_tests.rs
@@ -1,7 +1,10 @@
 use super::*;
 use crate::debate_cmd_output::*;
+use crate::debate_cmd_resolve::resolve_debate_tool;
 use csa_config::global::ReviewConfig;
+use csa_config::{GlobalConfig, ProjectConfig};
 use csa_config::{ProjectMeta, ResourcesConfig, ToolConfig};
+use csa_core::types::ToolName;
 use serde_json::Value;
 use std::collections::HashMap;
